# src-ui/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Rust Gallery API",
    "description": "API for managing photo galleries with folders and tags",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://192.168.1.27:8000",
      "description": "Main server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local rust server"
    }
  ],
  "paths": {
...
    "/folders/json/name/{folderName}": {
      "get": {
        "tags": [
          "folder"
        ],
        "summary": "Get folder by name",
        "description": "Returns details of a specific folder by its name",
        "operationId": "getFolderByName",
        "parameters": [
          {
            "name": "folderName",
            "in": "path",
            "required": true,
            "description": "Name of the folder to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Folder"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/folders/delete": {
      "post": {
        "tags": [
          "folder"
        ],
        "summary": "Delete a folder",
        "description": "Deletes a folder and returns the number of affected rows",
        "operationId": "deleteFolder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "string",
                    "description": "Name of the folder to delete"
                  }
                },
                "required": [
                  "data"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletion status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rows": {
                      "type": "integer",
                      "description": "Number of rows affected"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Folder": {
        "type": "object",
        "description": "Represents a folder in the gallery",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the folder"
          },
          "name": {
            "type": "string",
            "description": "Name of the folder"
          },
          "path": {
            "type": "string",
            "description": "File system path to the folder"
          },
          "thumbnail": {
            "type": "string",
            "description": "URL or path to the folder's thumbnail image"
          },
          "tags": {
            "type": "array",
            "description": "List of tags associated with the folder",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}